api/requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
Pillow
drf-extra-fields
PyPDF2